egl-wayland: Validate and use the passed value of EGL_EXT_present_opaque
It was assumed that the presense of EGL_PRESENT_OPAQUE_EXT in the surface
attribute list meant that the alpha channel should be ignored as if it was
always followed by a value of EGL_TRUE, however, per the spec, valid usage
allows for a value of EGL_FALSE as well, which should result in the alpha
channel being used for blending as normal.

Validate and use the value instead of assuming that the presense of the
extension enum value implies that it is enabled.